+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * `tunnel` dumped app argument
   
 * server
+  * moved fields from `app` to `status`: `container`, `uptime`
   * `start` replaced apps argument by optional image arguments; returns a single image now
   * `stop` dumped apps argument, returns an optional String now
   * `appInfo` dumped app argument